pyproject.toml
[project]
name = "un-website-boilerplate"
version = "0.1.0"
description = "UN Website Boilerplate"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "joblib>=1.4.0",
    "natsort>=8.4.0",
    "openai>=1.0.0",
    "pandas>=2.3.3",
    "pyairtable>=3.3.0",
    "pydantic>=2.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
